+++ azext_iot/operations/events3/_events.py
@@ -26,9 +26,9 @@ logger = get_logger(__name__)
 DEBUG = True
 
 
-def executor(target, consumer_group, enqueued_time, device_id=None, properties=None, timeout=0, output=None):
+def executor(target, consumer_group, enqueued_time, device_id=None, properties=None, timeout=0, output=None, content_type=None):
     coroutines = []
-    coroutines.append(initiate_event_monitor(target, consumer_group, enqueued_time, device_id, properties, timeout, output))
+    coroutines.append(initiate_event_monitor(target, consumer_group, enqueued_time, device_id, properties, timeout, output, content_type))
 
     loop = asyncio.get_event_loop()
     if loop.is_closed():
@@ -65,7 +65,7 @@ def executor(target, consumer_group, enqueued_time, device_id=None, properties=N
                 raise RuntimeError(error)
 
 
-async def initiate_event_monitor(target, consumer_group, enqueued_time, device_id=None, properties=None, timeout=0, output=None):
+async def initiate_event_monitor(target, consumer_group, enqueued_time, device_id=None, properties=None, timeout=0, output=None, content_type=None):
     def _get_conn_props():
         properties = {}
         properties["product"] = "az.cli.iot.extension"
@@ -108,12 +108,13 @@ async def initiate_event_monitor(target, consumer_group, enqueued_time, device_i
                                              properties=properties,
                                              device_id=device_id,
                                              timeout=timeout,
-                                             output=output))
+                                             output=output,
+                                             content_type=content_type))
         await asyncio.gather(*coroutines, return_exceptions=True)
 
 
 async def monitor_events(endpoint, connection, path, auth, partition, consumer_group, enqueuedtimeutc,
-                         properties, device_id=None, timeout=0, output=None):
+                         properties, device_id=None, timeout=0, output=None, content_type=None):
     source = uamqp.address.Source('amqps://{}/{}/ConsumerGroups/{}/Partitions/{}'.format(endpoint, path,
                                                                                          consumer_group, partition))
     source.set_filter(
@@ -134,11 +135,14 @@ async def monitor_events(endpoint, connection, path, auth, partition, consumer_g
         data = msg.get_data()
         if data:
             payload = str(next(data), 'utf8')
-        
-        system_props = unicode_binary_map(parse_entity(msg.properties, True))
-        content_type = system_props['content_type'].lower() if 'content_type' in system_props else ''
 
-        if content_type == 'application/json':
+        system_props = unicode_binary_map(parse_entity(msg.properties, True))
+
+        ct = content_type
+        if not ct:
+            ct = system_props['content_type'].lower() if 'content_type' in system_props else ''
+
+        if ct == 'application/json':
             try:
                 payload = json.loads(re.compile(r'(\\r\\n)+|\\r+|\\n+').sub('', payload))
             except Exception:  # pylint: disable=broad-except
@@ -160,7 +164,7 @@ async def monitor_events(endpoint, connection, path, auth, partition, consumer_g
 
             if app_prop:
                 event_source['event']['properties']['application'] = unicode_binary_map(app_prop)
-            
+
         if output.lower() == 'json':
             dump = json.dumps(event_source, indent=4)
         else: