+++ HISTORY.rst
@@ -7,7 +7,7 @@ Release History
 +++++++++++++++
 * Added --output support to monitor-events. Supports either json or yaml, i.e. az iot hub monitor-events --hub-name {} -d {} --output yaml
 * Changed monitor-events to output JSON by default
-* Added support to parse payload as JSON if system property Content-Type is provided and application/json (i.e. --props $.ct=application/json from the CLI)
+* Added support to parse and display payload as JSON if system property Content-Type is provided and application/json (i.e. send-d2c-message ... --props $.ct=application/json from the CLI) or if monitor-events has a property --content-type/--ct of application/json (i.e. monitor-events --ct application/json).
 
 0.6.0
 +++++++++++++++